Akylai/2 commits in 1
fix/changes commit

fix/changes in tests commit
